Handle duplicate DDU profiles
This basically runs a check at the beginning of a setup and sees if a user "DDU"'s folder exists, and if it does then it adds a "U" at the end of the one it's going to use until it finds an unoccupied space (or it reaches the 256 character cap, at which point it errors out).

This is a very ugly solution to working around this problem, and I simply chose this as a temporary option because I do not feel it is safe to delete profiles.