Fix verification of driver signature
Due to an oversight in how I thought Windows built-in Get-AuthenticodeSignature function worked, AutoDDU did not actually verify that the file had a proper signature, just that it was signed, but did not verify that the signature actually belonged to the file in question.

This now uses signify to actually verify the integrity of the file with the signature, and makes sure the digest hash match the expected.